Create moduleswitcher.py and update main.py
moduleswitcher.py will switch to different modules based on the input speech queried with the different lists and then switch to that module and perform the specific operation.
Added the searchengine module in the switcher.
Any number of modules can be added by adding a new list in moduleswitcher.py and comparing it with voice_in.lower() in the elif statements.
Removed shifter function from main.py to moduleswitcher.